requirements.txt
numpy==2.4.4
pandas==3.0.2
scikit-learn==1.8.0
fastapi[all]==0.136.0
uvicorn==0.24.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.0.1
httpx==0.27.0
pytest==7.2.1
pycodestyle==2.10.0
python-keycloak==5.0.0
PyJWT==2.8.0